# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "S.A.R.A. API",
        "description": "API para el Sistema Ágil de Rutas y Asignaciones",
        "contact": {
            "email": "support@sara.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pedidos": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Listar pedidos",
                "operationId": "9368500233e75a120cb606e08d7a06e4",
                "responses": {
                    "200": {
                        "description": "Lista de pedidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "cliente_id": {
                                                "type": "integer"
                                            },
                                            "conductor_id": {
                                                "type": "integer",
                                                "nullable": true
                                            },
                                            "origen": {
                                                "type": "string"
                                            },
                                            "destino": {
                                                "type": "string"
                                            },
                                            "descripcion": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "estado": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Crear un nuevo pedido",
                "operationId": "acf9c44ffbceb280fd290ecbb5f5f3f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "origen",
                                    "destino"
                                ],
                                "properties": {
                                    "origen": {
                                        "type": "string",
                                        "example": "Ciudad A"
                                    },
                                    "destino": {
                                        "type": "string",
                                        "example": "Ciudad B"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Carga de 2 toneladas",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "cliente_id": {
                                            "type": "integer"
                                        },
                                        "origen": {
                                            "type": "string"
                                        },
                                        "destino": {
                                            "type": "string"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "estado": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pedidos/{id}": {
            "put": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Actualizar un pedido",
                "operationId": "3da8de6967b1e89d1d5a0cf5988c04e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "origen": {
                                        "type": "string",
                                        "example": "Ciudad A"
                                    },
                                    "destino": {
                                        "type": "string",
                                        "example": "Ciudad B"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Carga de 2 toneladas",
                                        "nullable": true
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "pendiente",
                                            "en_progreso",
                                            "completado"
                                        ],
                                        "example": "en_progreso"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pedido actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "cliente_id": {
                                            "type": "integer"
                                        },
                                        "origen": {
                                            "type": "string"
                                        },
                                        "destino": {
                                            "type": "string"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "estado": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Eliminar un pedido",
                "operationId": "c0cc39e3f79740c3fe58d2268a321d0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pedido eliminado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pedido-conductor": {
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Asignar conductor a un pedido",
                "operationId": "3a323f54d131695947ba883283c1a20a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pedido_id",
                                    "conductor_id"
                                ],
                                "properties": {
                                    "pedido_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "conductor_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Conductor asignado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "cliente_id": {
                                            "type": "integer"
                                        },
                                        "conductor_id": {
                                            "type": "integer"
                                        },
                                        "origen": {
                                            "type": "string"
                                        },
                                        "destino": {
                                            "type": "string"
                                        },
                                        "estado": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}